Seperation between service and repository
SponaacularAPIrepository has now the functions from service before
