+++ src/main/java/com/RecipeFinder/backend/controllers/RecipeController.java
@@ -3,6 +3,7 @@ package com.RecipeFinder.backend.controllers;
 import com.RecipeFinder.backend.services.RecipeService;
 import com.RecipeFinder.backend.models.Recipe;
 import com.RecipeFinder.backend.repositories.RecipeRepository;
+import com.RecipeFinder.backend.repositories.SpoonacularAPIRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
