+++ src/main/java/com/RecipeFinder/backend/services/RecipeService.java
@@ -1,75 +1,24 @@
 package com.RecipeFinder.backend.services;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
-
-import com.RecipeFinder.backend.models.Recipe;
-import com.RecipeFinder.backend.repositories.RecipeRepository;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.util.ArrayList;
+import com.RecipeFinder.backend.repositories.SpoonacularAPIRepository;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class RecipeService {
-    private final String apiUrl = "https://api.spoonacular.com/recipes/complexSearch";
-    private final String apiKey = "4b2cdc4ff73546b89cc40882c81c8e9c";
-    
-  private List<Integer> recipeIds = new ArrayList<>();
-  
-  public void applyIngredientFilter(List<String> ingredients) {
-    RestTemplate restTemplate = new RestTemplate();
-    //Creating comma separated string
-    String includeIngrdients = ingredients.stream().collect(Collectors.joining(","));
-    String url = apiUrl + "?apiKey=" + apiKey + "&query=pasta" + "&number=3" + "&includeIngredients=" + includeIngrdients;
+    @Autowired
+    private SpoonacularAPIRepository SpoonacularAPIRepository;
 
-    try {
-      String response = restTemplate.getForObject(url, String.class);
-      System.out.println("API Response: " + response);
-
-      //Parsing
-      ObjectMapper mapper = new ObjectMapper();
-      JsonNode root = mapper.readTree(response);
-      JsonNode results = root.path("results");
-
-      recipeIds.clear();
-
-      for (JsonNode result : results) {
-        int id = result.path("id").asInt();
-        recipeIds.add(id);
-        System.out.println("Recipe ID: " + id);
-      }
-
-    } catch (Exception e) {
-            e.printStackTrace();
-        }
-  }
-
-  private final String infoUrl = "https://api.spoonacular.com/recipes/{id}/information";
-  public void getRecipeInformation(Integer id) {
-    RestTemplate restTemplate = new RestTemplate();
-    try {
-      String url = infoUrl.replace("{id}", id.toString()) + "?apiKey=" + apiKey;
-      String response = restTemplate.getForObject(url, String.class);
-      System.out.println("Recipe Information for ID " + id + ": " + response);
-    } catch (Exception e) {
-      e.printStackTrace();
+    // Service method to the recipes with the ingredients
+    public void applyIngredientFilter(List<String> ingredients) {
+      SpoonacularAPIRepository.applyIngredientFilter(ingredients);
     }
 
-  }
-
-  public String extractItemName(String item) {
-    String[] parts = item.split(" ", 2);
-    //For error handling
-    if (parts.length > 1) {
-      return parts[1].toLowerCase();
+    // Service method to get recipe informatio with recipe id
+    public void getRecipeInformation(Integer id) {
+      SpoonacularAPIRepository.getRecipeInformation(id);
     }
-    return "";
-  }
-
-
-
 }
+
+
